+++ kmjxc/dba/Back_Stock.cs
@@ -15,12 +15,12 @@ namespace KM.JXC.DBA
     public partial class Back_Stock
     {
         public int Back_Sock_ID { get; set; }
-        public int Back_Date { get; set; }
+        public long Back_Date { get; set; }
         public int Shop_ID { get; set; }
         public int User_ID { get; set; }
         public string Description { get; set; }
         public int Back_Sale_ID { get; set; }
-        public int Created { get; set; }
+        public long Created { get; set; }
         public int Status { get; set; }
     }
 }
