+++ kmjxc/dba/Express_Shop.cs
@@ -17,9 +17,9 @@ namespace KM.JXC.DBA
         public int Express_ID { get; set; }
         public int Shop_ID { get; set; }
         public int IsDefault { get; set; }
-        public int Created { get; set; }
+        public long Created { get; set; }
         public int Created_By { get; set; }
-        public int Modified { get; set; }
+        public long Modified { get; set; }
         public int Modified_By { get; set; }
     }
 }
