+++ kmjxc/dba/Sale_SyncTime.cs
@@ -15,9 +15,9 @@ namespace KM.JXC.DBA
     public partial class Sale_SyncTime
     {
         public int ShopID { get; set; }
-        public int LastSyncTime { get; set; }
-        public int LastTradeStartEndTime { get; set; }
-        public int LastTradeModifiedEndTime { get; set; }
+        public long LastSyncTime { get; set; }
+        public long LastTradeStartEndTime { get; set; }
+        public long LastTradeModifiedEndTime { get; set; }
         public int SyncUser { get; set; }
         public int SyncType { get; set; }
     }
